pyproject.toml
[tool.poetry]
name = "true-test"
version = "0.1.0"
description = ""
authors = ["parkkk123 <tg-gmnews@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
jupyter = "^1.0.0"
scikit-surprise = "^1.1.4"
numpy = "1.26.0"
fastapi = "^0.111.1"
pytest = "^8.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
